requirements.txt
Django==1.10.4
Pillow
psycopg2
django-crispy-forms==1.4.0
python-dateutil